moved avoidance_target_detector to a proper place.
