+++ planning/autoware_diffusion_planner/CMakeLists.txt
@@ -80,7 +80,7 @@ if(TRT_AVAIL AND CUDA_AVAIL)
   set(DIFFUSION_PLANNER_SOURCES
     src/diffusion_planner_core.cpp
     src/diffusion_planner_node.cpp
-    src/avoidance_target_filter.cpp
+    src/mppi_related/avoidance_target_filter.cpp
     src/conversion/lanelet.cpp
     src/conversion/agent.cpp
     src/postprocessing/postprocessing_utils.cpp
